adding routes (+ updating CSS to match)
