+++ react-ui/src/components/ViewInterface/ViewInterface.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+function ViewInterface(props) {
+  return (
+    <h1>View</h1>
+  );
+}
+
+export default ViewInterface;
