+++ react-ui/src/index.css
@@ -55,11 +55,16 @@ button:active{
 }
 
 .App {
+  margin: 0;
+  padding: 0;
+}
+
+.main-section
+{
   display: flex;
   flex-direction: column;
   align-items: center;
-  margin: 0;
-  padding: 0;
+  width: 100%;
 }
 
 .control-panel {
@@ -120,25 +125,24 @@ button:active{
 
 @media (min-width: 1020px) {
   .App {
-    flex-direction: row;
     margin: 0;
     padding: 0;
     height: 100%;
     min-height: 100%;
-    align-items: flex-start;
-    justify-content: flex-start;
   }
-  .top-section {
+  .main-section {
     height: 100%;
     min-height: 100vh;
-    width: 70%;
+    flex-direction: row;
+    align-items: flex-start;
+    justify-content: space-around;
   }
   .control-panel {
     height: 100%;
-    margin: 0;
     min-height: 100vh;
-    width: 30%;
     max-height: 100vh;
+    width: 30%;
+    margin: 0;
     overflow-y: scroll;
     scrollbar-color: white;
   }
